M10Clone/backend/requirements_telegram.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
python-telegram-bot==20.7
